+++ aesara/scan/op.py
@@ -70,9 +70,9 @@ from aesara.graph.basic import (
     graph_inputs,
     io_connection_pattern,
 )
+from aesara.graph.features import NoOutputFromInplace
 from aesara.graph.fg import MissingInputError
 from aesara.graph.op import Op, ops_with_inner_function
-from aesara.graph.toolbox import NoOutputFromInplace
 from aesara.link.c.basic import CLinker
 from aesara.link.c.exceptions import MissingGXX
 from aesara.link.utils import raise_with_op
