+++ aesara/tensor/blas.py
@@ -146,6 +146,7 @@ import aesara.scalar
 from aesara.compile.mode import optdb
 from aesara.configdefaults import config
 from aesara.graph.basic import Apply, view_roots
+from aesara.graph.features import ReplacementDidNotRemoveError, ReplaceValidate
 from aesara.graph.fg import InconsistencyError
 from aesara.graph.op import COp, Op
 from aesara.graph.opt import (
@@ -157,7 +158,6 @@ from aesara.graph.opt import (
 )
 from aesara.graph.optdb import SequenceDB
 from aesara.graph.params_type import ParamsType
-from aesara.graph.toolbox import ReplacementDidNotRemoveError, ReplaceValidate
 from aesara.graph.utils import MethodNotDefined, TestValueError
 from aesara.printing import FunctionPrinter, debugprint, pprint
 from aesara.scalar import bool as bool_t
