+++ setup.cfg
@@ -4,6 +4,7 @@ ignore = E203,E231,E501,E741,W503,W504,C901
 max-line-length = 88
 per-file-ignores =
     **/__init__.py:F401,E402,F403
+    aesara/graph/toolbox.py:E402,F403,F401
     aesara/link/jax/jax_dispatch.py:E402,F403,F401
     aesara/link/jax/jax_linker.py:E402,F403,F401
     aesara/sparse/sandbox/sp2.py:F401
